+++ src/app/entities/LivreEmpruntes.java
@@ -1,4 +1,4 @@
-package app.models;
+package app.entities;
 
 import java.util.Date;
